Refactoring of the Privacy Protection screen
