+++ DuckDuckGo/UserText.swift
@@ -73,6 +73,9 @@ public struct UserText {
     
     public static let navigationTitleEdit = NSLocalizedString("navigation.title.edit", value: "Edit", comment: "Edit button")
     
+    public static let privacyProtectionProtectionDisabled = NSLocalizedString("privacy.protection.main.disabled", comment: "SITE PROTECTION DISABLED")
+    public static let privacyProtectionPrivacyGrade = NSLocalizedString("privacy.protection.main.grade", comment: "PRIVACY GRADE")
+    public static let privacyProtectionEnhanced = NSLocalizedString("privacy.protection.main.enhanced", comment: "ENHANCED FROM $1 TO $2")
     public static let privacyProtectionTrackersBlocked = NSLocalizedString("privacy.protection.trackers.blocked", comment: "Trackers blocked")
     public static let privacyProtectionTrackersFound = NSLocalizedString("privacy.protection.trackers.found", comment: "Trackers found")
     public static let privacyProtectionMajorTrackersBlocked = NSLocalizedString("privacy.protection.major.trackers.blocked", comment: "Major trackers blocked")
